requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
mss==10.1.0
Pillow==11.3.0
python-dotenv==1.1.1
pytesseract==0.3.13
